time: Fix timespec_to_nsec()
Convert to uint64_t before doing the conversion to eliminate
overflow.

Reviewed-on: http://gerrit.mot.com/836818
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>

+++ nuttx/include/nuttx/time.h
@@ -152,7 +152,7 @@ static inline useconds_t timespec_to_usec(const struct timespec *ts)
  ****************************************************************************/
 static inline uint64_t timespec_to_nsec(const struct timespec *ts)
 {
-	return ts->tv_sec * NSEC_PER_SEC + ts->tv_nsec;
+	return ((uint64_t)ts->tv_sec) * NSEC_PER_SEC + ts->tv_nsec;
 }
 
 /****************************************************************************
